back/requirements.txt
torch
torchvision
flask
flask-cors
pillow
scikit-image